Organized lobby related functions into different files
