+++ server/src/lobby/client_handler.rs
@@ -1,78 +1,24 @@
+use crate::Client;
+use super::util::{
+    Message::{self, BoardState, SpecialMessage},
+    Subtasks
+};
+
 use constants::ConnectionProtocol;
 
 use tokio::{
-    net::TcpStream,
     sync::{
-        broadcast::{self, Sender, Receiver},
-        mpsc::{self, UnboundedSender, UnboundedReceiver}
+        broadcast::{Sender, Receiver},
+        mpsc::{UnboundedSender, UnboundedReceiver}
     },
-    task::{self, JoinHandle}
+    task
 };
-use tokio_tungstenite::{
-    tungstenite::Message::Binary,
-    WebSocketStream
-};
+use tokio_tungstenite::tungstenite::Message::Binary;
 use futures::{SinkExt, StreamExt, stream::{SplitSink, SplitStream}};
 
-use std::{
-    collections::HashMap,
-    sync::{Arc, Mutex}
-};
+use std::sync::{Arc, Mutex};
 
-pub type Client = WebSocketStream<TcpStream>;
-pub type Lobbies = HashMap<String, UnboundedSender<Client>>;
-
-struct Subtasks {
-    tasks: Vec<JoinHandle<()>>,
-    last_board_state: u64
-}
-
-impl Default for Subtasks {
-    fn default() -> Self {
-        Subtasks { tasks: Vec::new(), last_board_state: 0 }
-    }
-}
-
-#[derive(Debug)]
-enum Message {
-    BoardState(u64),
-    SpecialMessage(u8)
-}
-
-use Message::{BoardState, SpecialMessage};
-
-async fn run_lobby(
-    mut receiver: UnboundedReceiver<Message>,
-    game_update_sender: Sender<Vec<u8>>,
-    subtasks: Arc<Mutex<Subtasks>>,
-    remove_lobby: Box<dyn FnOnce() -> () + Send + Sync>
-) {
-
-    while let Some(msg) = receiver.recv().await {
-        match msg {
-
-            BoardState(state) => {
-                { subtasks.lock().unwrap().last_board_state = state; }
-                game_update_sender.send(ConnectionProtocol::disassemble_message(state)).unwrap_or_default();
-            }
-            SpecialMessage(_) => {
-                break;
-            }
-
-        }
-    }
-
-    task::block_in_place(move || {
-        remove_lobby();
-
-        for subtask in &subtasks.lock().unwrap().tasks {
-            subtask.abort();
-        }
-    });
-    println!("Ending lobby.");
-}
-
-async fn new_client_handler(
+pub async fn new_client_handler(
     sender: UnboundedSender<Message>,
     mut new_client_receiver: UnboundedReceiver<Client>,
     game_update_sender: Sender<Vec<u8>>,
@@ -139,27 +85,4 @@ async fn client_writer(mut client: SplitSink<Client, tokio_tungstenite::tungsten
             break;
         }
     }
-}
-
-pub fn create_lobby(remove_lobby: Box<dyn FnOnce() -> () + Send + Sync>) -> UnboundedSender<Client> {
-
-    let (sender, receiver) = mpsc::unbounded_channel();
-    let (new_client_sender, new_client_receiver)
-        = mpsc::unbounded_channel();
-
-    let (game_update_sender, _) = broadcast::channel(3);
-    let game_update_sender_clone = game_update_sender.clone();
-
-    let subtasks = Arc::new(Mutex::new(Subtasks::default()));
-    let subtasks_ref = Arc::clone(&subtasks);
-
-    task::spawn(async move {
-        run_lobby(receiver, game_update_sender, subtasks, remove_lobby).await;
-    });
-    task::spawn(async move {
-        new_client_handler(sender, new_client_receiver, game_update_sender_clone, subtasks_ref).await;
-    });
-    
-    new_client_sender
-
 }
